% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ml.R
\name{create_ml}
\alias{create_ml}
\title{Create a machine learning classifier}
\usage{
create_ml(data_source, method)
}
\arguments{
\item{data_source}{The data to use to build the classifier.
Either a data frame, or a file name or a URL. The \code{rio}
package is used to read the data from files and URLs, and
the data is converted to a data frame after reading it.}

\item{method}{Name of the method to use. Currently \code{knn},
the \eqn{k} nearest neighbors classifier, and \code{lm},
linear models, are implemented.}
}
\description{
Create a machine learning classifier
}

