% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/package.R
\docType{package}
\name{ rclasses.S3 }
\alias{ rclasses.S3 }
\alias{ rclasses.S3 }
\title{ Example Package for S3 Classes }
\description{
R supports three systems for writing object orineted code, and some contributed pakages add even more. This package is part of a systematic comparison of all class systems, by implementing a simple generic machine learning framework.
}
